完善歌单页面功能

+++ app/src/main/java/me/rerere/rainmusic/MainActivity.kt
@@ -121,7 +121,7 @@ class MainActivity : ComponentActivity() {
                                     }
                                 )
                             ) {
-                                PlaylistScreen(id = it.arguments!!.getInt("id"))
+                                PlaylistScreen(id = it.arguments!!.getLong("id"))
                             }
                         }
                     }
